requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils
psycopg2-binary
alembic
redis
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
black
isort
pydantic[email]